+++ R/VarImp.R
@@ -2,7 +2,7 @@
 #'
 #' Variable importance is computed from permuting OOB data.
 #'
-#' @param forest An object of class \code{\link{ODRF}}.
+#' @param obj An object of class \code{\link{ODRF}}.
 #' @param X An n by d numerical matrix (preferably) or data frame is used in the \code{ODRF}.
 #' @param y A response vector of length n is used in the \code{ODRF}.
 #'
@@ -19,14 +19,13 @@
 #' set.seed(221212)
 #' train <- sample(1:569, 200)
 #' train_data <- data.frame(breast_cancer[train, -1])
-#' test_data <- data.frame(breast_cancer[-train, -1])
-#'\donttest{
-#' forest <- ODRF(diagnosis ~ ., train_data, split = "gini", parallel = FALSE)
-#' (varimp <- VarImp(forest, train_data[, -1], train_data[, 1]))
-#'}
+#' forest <- ODRF(train_data[, -1], train_data[, 1], split = "gini",
+#'   parallel = FALSE)
+#' varimp <- VarImp(forest, train_data[, -1], train_data[, 1])
+#' varimp
 #' @keywords forest
 #' @export
-VarImp <- function(forest, X, y) {
+VarImp <- function(obj, X, y) {
   # vars=all.vars(forest$terms)
   # address na values.
   # if (any(is.na(data))) {
@@ -35,7 +34,19 @@ VarImp <- function(forest, X, y) {
   # }
   # y= data[,setdiff(colnames(data),vars[-1])]
   # X= data[,vars[-1]]
+
+  forest=obj
   X <- as.matrix(X)
+  colnames(X)=forest$data$varName
+
+  pp <- forest$data$p
+  if (!is.null(forest$data$catLabel) && (sum(forest$data$Xcat) > 0)) {
+    pp <- pp - length(unlist(forest$data$catLabel)) + length(forest$data$Xcat)
+  }
+  if (ncol(X) != pp) {
+    stop("The dimensions of 'Xnew' and training data do not match")
+  }
+
 
   if (!is.null(forest$data$subset)) {
     X <- X[forest$data$subset, ]
@@ -46,7 +57,8 @@ VarImp <- function(forest, X, y) {
 
 
   if (forest$split != "mse") {
-    y <- factor(y, levels = forest$Levels)
+    #y <- factor(y, levels = forest$Levels)
+    y <- as.character(y)
   }
   # X=forest$data$X
   # y=forest$data$y
@@ -81,6 +93,9 @@ VarImp <- function(forest, X, y) {
     rm(X1)
     rm(Xj)
   }
+  if (!is.numeric(X)){
+    X=apply(X, 2, as.numeric)
+  }
 
   # Variable scaling.
   if (forest$data$Xscale != "No") {
@@ -91,42 +106,45 @@ VarImp <- function(forest, X, y) {
 
   runOOBErr <- function(tree, ...) {
     class(tree) <- "ODT"
-    oobErrs <- rep(0, p + 1)
+    oobErrs <- rep(0, p)
     oobIndex <- tree$oobIndex
-
-    Xi <- X[oobIndex, ]
-    yi <- y[oobIndex]
-    yn <- length(yi)
+    X0 <- X[oobIndex, ]
+    y0 <- y[oobIndex]
+    n0 <- length(y0)
     # if(forest$split=="regression"){
     #  e.0=mean((yi-mean(y[-oobIndex]))^2)
     # }
-    for (j in 1:(p + 1)) {
-      if (j != 1) {
-        Xi[, j - 1] <- Xi[sample(yn), j - 1] #+rnorm(length(oobIndex))
-      }
+
+    pred <- predict(tree, X0)
+    if (forest$split != "mse") {
+      oobErr0 <- mean(pred != y0)
+    } else {
+      oobErr0 <- mean((pred - y0)^2) # /e.0
+    }
+
+    for (j in seq(p)) {
+      Xi=X0
+      Xi[, j] <- Xi[sample(n0), j] #+rnorm(length(oobIndex))
       pred <- predict(tree, Xi)
       if (forest$split != "mse") {
-        oobErr <- mean(pred != yi)
+        oobErr <- mean(pred != y0)
       } else {
-        oobErr <- mean((pred - yi)^2) # /e.0
+        oobErr <- mean((pred - y0)^2) # /e.0
       }
 
-      if (j == 1) {
-        oobErrs[1] <- oobErr
-      } else {
-        oobErrs[j] <- abs(oobErr - oobErrs[1])
-      }
+      oobErrs[j] <- abs(oobErr - oobErr0)
     }
 
     return(oobErrs)
   }
 
-  oobErrVar <- sapply(forest$ppTrees, runOOBErr)[-1, ]
-  oobErrVar <- rowMeans(oobErrVar)
-  varimport <- cbind(varible = seq(p), increased_error = oobErrVar)
+  IncErr <- vapply(forest$ppTrees, runOOBErr, rep(0,p))
+  #IncErr <- sapply(forest$ppTrees, runOOBErr)
+  IncErr <- rowMeans(IncErr)
+  varimport <- cbind(varible = seq(p), increased_error = IncErr)
   rownames(varimport) <- colnames(X)
 
-  varimport <- list(varImp = varimport[order(oobErrVar, decreasing = T), ], split = forest$split)
+  varimport <- list(varImp = varimport[order(IncErr, decreasing = T), ], split = forest$split)
   class(varimport) <- "VarImp"
 
   return(varimport)