+++ R/plot_VarImp.R
@@ -17,22 +17,18 @@
 #' set.seed(221212)
 #' train <- sample(1:569, 200)
 #' train_data <- data.frame(breast_cancer[train, -1])
-#' test_data <- data.frame(breast_cancer[-train, -1])
-#'\donttest{
-#' forest <- ODRF(diagnosis ~ ., train_data, split = "entropy", parallel = FALSE)
-#' (varimp <- VarImp(forest, train_data[, -1], train_data[, 1]))
-#' plot(varimp, digits = 0)
-#'}
+#' forest <- ODRF(train_data[, -1], train_data[, 1], split = "gini",
+#'   parallel = FALSE)
+#' varimp <- VarImp(forest, train_data[, -1], train_data[, 1])
+#' plot(varimp)
 #' @keywords forest plot
 #' @rdname plot.VarImp
 #' @aliases plot.VarImp
 #' @method plot VarImp
 #' @export
-plot.VarImp <- function(x, nvar = 30, digits = NULL, main = NULL, ...) {
+plot.VarImp <- function(x, nvar = min(30,nrow(x$varImp)), digits = NULL, main = NULL, ...) {
   imp <- x$varImp
   imp <- imp[1:nvar, , drop = FALSE]
-
-  nvar <- min(nvar, nrow(x$varImp))
 
   if (is.null(main)) {
     main <- paste0("Oblique ", ifelse(x$split == "mse", "Regression", "Classification"), " Forest")
@@ -50,11 +46,20 @@ plot.VarImp <- function(x, nvar = 30, digits = NULL, main = NULL, ...) {
     }
   }
 
+  if(digits==0){
+    xlab = paste0("Increased ",ifelse(x$split == "mse","MSE","MR"))
+  }else if(digits==2){
+    xlab = paste0("Increased ",ifelse(x$split == "mse","MSE","MR")," (%)")
+  }else{
+    xlab = substitute(paste("Increased ",ifelse(x$split == "mse","MSE","MR")," (*",10^{-dig},")"),list(dig = digits))
+  }
 
   ## If there are more than two columns, just use the last two columns.
   #op <- par(xaxs = "i") #* 10^digits
+
+
   dotchart(sort(imp[, 2]),
-    xlab = paste0("Increased error (*", 10^-digits, ")"), ylab = "", main = main, xaxt = "n",
+    xlab = xlab, ylab = "", main = main, xaxt = "n",
     cex.lab = 1.5, cex.axis = 1.25, bg = "skyblue"
   )
   axis(1, seq(min(imp[, 2]), max(imp[, 2]), length.out = min(6, nvar)),